🚚 Export everything and fix jsdocs
